// frontend/src/types/supabase.ts
export type UserRole = 
  | 'tier1' 
  | 'tier2_officer' 
  | 'tier2_registrar' 
  | 'company_admin' 
  | 'company_manager' 
  | 'company_user' 
  | 'auditor' 
  | 'system_admin' 
  | 'vendor'